Move composeMosaicName fn to mosaic utils
